Create EditSubCalendarPermissions.ps1
When working at the helpdesk, I had a few requests to add users to the permissions list for a shared mailbox/calendar. I found there was a powershell command to do it, instead of adding myself as a delegate to the mailbox, but the process was a little weird. I ended up writing a script to do it with a WPF interface.